+++ src/main/kotlin/toliner/nbt/CompoundMap.kt
@@ -8,7 +8,6 @@
 
 package toliner.nbt
 
-import toliner.nbt.tags.NBTTag
 import java.util.*
 import kotlin.collections.LinkedHashMap
