+++ src/main/kotlin/toliner/nbt/TagType.kt
@@ -8,7 +8,6 @@
 
 package toliner.nbt
 
-import toliner.nbt.tags.NBTTag
 import kotlin.reflect.KClass
 
 /**
